Tint the extension's toolbar icon in dev and staging environments
Tint the extension icon red-orange in dev builds
and yellow-gold in staging builds to make it easier
to tell the different builds apart.

This commit programatically generates tinted images
based on the original icons. The advantage is that this
approach can be used for as many variants as we like and
possibly animations or other effects on the badge image.

However, it might just be simpler to variations on the original
assets for dev and staging builds.