Replace the CSS spinner with a Material Design-style spinner
This updates the loading spinner displayed when fetching
annotations and performing other network requests
to a more modern design that also avoids the blurry
rendering seen in Firefox and Chrome with the previous
implementation.

The new spinner is implemented using rAF + <canvas>
and brings in one small dependency for the
cubic bezier timing function.

T-139