Remove Angular from webassets
Use webpack to build and bundle Angular.
It is currently built as part of the main app bundle,
in future we can use Webpack's code-splitting facilities
to separate out vendor code into a separate bundle.

The 'config/module' module also has an Angular dependency
and needs to be run after the main `h` Angular module
has been declared, so move it into a CommonJS require
in app.coffee