Avoid requiring access to all URLs on initial installation
When the extension is initially installed, avoid asking
for access to all URLs but instead just ask for the 'activeTab'
and 'tabs' permissions. This results in a less scary prompt
which asks for access to the user's browsing history but not
access to their data on all sites.

If the user navigates whilst H is active on a tab, the extension
will show the settings dialog at that point to prompt
for additional permissions.