Add a more accurate mechanism to detect when the sidebar is unloaded
Previously the extension detected that the sidebar was unloaded
by listening for chrome.tabs.onUpdated to be emitted with a 'loading'
status. However, this reports false positives since onUpdated is emitted
when subframes within a tab enter a loading state, which does not
affect the sidebar.

Instead, add a listener for the top-level frame's "unload" event
via a content script which sends a message back to the extension's
background script notifying it that the tab has been unloaded.

The background script then updates its internal state indicating that
the sidebar is no longer installed.