Experiment with using chrome.webNavigation to detect page navigation instead of chrome.tabs
chrome.webNavigation provides finer grained notification of tab
load states than the chrome.tabs API, enabling earlier injection
of the sidebar when navigating from one page to another with
the extension enabled via the DOMContentLoaded notification.

More importantly, it can distinguish between navigations
in top-level browsing contexts vs iframes.

A downside is that for pre-rendered tabs, it can report
events for tab processes that are not yet associated
with visible tabs in the browser but that later become
visible via a chrome.webNavigation.onTabReplaced event.