Avoid trying to set properties associated with hidden tabs
The chrome.webNavigation API reports events for pre-rendered tabs
which are hidden and not associated with any visible tab.

These tabs are not accessible to the chrome.tabs or
chrome.browserAction APIs so ignore them. The state for such tabs
will be updated in the chrome.tabs.onReplaced handler when the tab
is swapped in.