Improve PDF viewer detection in Chrome
Rather than relying on the tab URL containing '.pdf',
check whether the current tab is displaying Chrome's native
PDF viewer and open Hypothesis' PDF.js based viewer in that
case since we the sidebar cannot be injected into the native viewer.

We detect the Chrome PDF viewer by checking for an
<embed ... type="application/pdf"> tag which appears at the top
of the body.

If the user is using a PDF.js or other HTML-based viewer,
then Hypothesis can work with that directly.

An alternative approach would be to detect the mime type of the
loaded document but that requires intercepting all page
load requests in order to read the Content-Type header, which
is not available after the page has loaded OR firing a network
request.

See also this upstream issue: https://code.google.com/p/chromium/issues/detail?id=242575