Refactor sidebar injector test
Instead of creating the injector in the beforeEach() function
and then customizing the result of content scripts by modifying
a couple of test-global variables, instantiate a new injector
directly in each test, which uses test-specific stubs set
up by a particular test.

This reduces the amount of global state in the test suite.