Reload whole page after changing username
By default when inline editing is enabled only the form is reloaded from
the returned HTML after submission.

In the case of a username change, reload the whole page in order to
refresh references to the username in the page's header after a
successful submit.

This is done via an extension to the inline form editing controller
rather than changing the username form to not use inline editing because
we want to preserve the other behaviors of inline editing.