Use sequence schema for "origins" field of group creation schema
Now that we have support for rendering list inputs in forms (albeit in
an unpolished way), use colander's standard method of representing list
fields in schemas for group origins.

This enables validation of the content of each list item using a
validator plus whatever UI affordances we have in our forms for editing
lists of things.